Refactor Profile into ParentProfile and ChildProfile
Split the Profile component into separate ParentProfile and ChildProfile components for better separation of concerns and maintainability. The Profile component now delegates rendering based on the user's role. Added new files for child and parent profile views.
